+++ apps/frontend/src/app/employee-management/admin/page.tsx
@@ -1,4 +1,4 @@
-import Layout from "@/app/employee-management/layout";
+import Layout from "@/app/layout-sidebar";
 
 export default function Home() {
   return <Layout />
