+++ apps/frontend/src/components/ui/sidebar.tsx
@@ -497,7 +497,7 @@ const sidebarMenuButtonVariants = cva(
 
 function SidebarMenuButton({
   asChild = false,
-  isActive = false,
+  // isActive = false,
   variant = "default",
   size = "default",
   tooltip,
@@ -516,7 +516,7 @@ function SidebarMenuButton({
       data-slot="sidebar-menu-button"
       data-sidebar="menu-button"
       data-size={size}
-      data-active={isActive}
+      // data-active={isActive}
       className={cn(sidebarMenuButtonVariants({ variant, size }), className)}
       {...props}
     />
